Add separate "Exhaustive" tests that are excluded from Valgrind
