+++ CMakeLists.txt
@@ -170,8 +170,10 @@ set(OTHER_SOURCE_FILES
     src/apps/testapps/testH3Api.c
     src/apps/testapps/testH3SetToLinkedGeo.c
     src/apps/testapps/testH3ToLocalIj.c
+    src/apps/testapps/testH3ToLocalIjExhaustive.c
     src/apps/testapps/testH3Distance.c
     src/apps/testapps/testH3Line.c
+    src/apps/testapps/testH3LineExhaustive.c
     src/apps/testapps/testCoordIj.c
     src/apps/testapps/testCoordIjk.c
     src/apps/miscapps/h3ToGeoBoundaryHier.c
@@ -521,6 +523,13 @@ if(BUILD_TESTING)
     add_h3_test_with_arg(testH3NeighborRotations src/apps/testapps/testH3NeighborRotations.c 0)
     add_h3_test_with_arg(testH3NeighborRotations src/apps/testapps/testH3NeighborRotations.c 1)
     add_h3_test_with_arg(testH3NeighborRotations src/apps/testapps/testH3NeighborRotations.c 2)
+
+    # The "Exhaustive" part of the test name is used by the Valgrind CI test to exclude
+    # these files in .travis.yml.
+    add_h3_test(testH3ToLocalIjExhaustive src/apps/testapps/testH3ToLocalIjExhaustive.c)
+    add_h3_test(testH3LineExhaustive src/apps/testapps/testH3LineExhaustive.c)
+
+    add_custom_target(test-fast COMMAND ctest -E Exhaustive)
 endif()
 
 if(BUILD_BENCHMARKS)
